+++ packages/core/src/rules/oas3/no-server-variables-empty-enum.ts
@@ -6,7 +6,7 @@ enum enumError {
   invalidDefaultValue = 'invalidDefaultValue',
 }
 
-export const NoEmptyEnumServers: Oas3Rule = () => {
+export const NoServerVariablesEmptyEnum: Oas3Rule = () => {
   return {
     DefinitionRoot(root, { report, location }) {
       if (!root.servers || root.servers.length === 0) return;
@@ -48,7 +48,7 @@ function checkEnumVariables(server: Oas3Server): enumError[] | undefined {
   if (server.variables && Object.keys(server.variables).length === 0) return;
 
   const errors: enumError[] = [];
-  for (var variable in server.variables) {
+  for (const variable in server.variables) {
     const serverVariable = server.variables[variable];
     if (!serverVariable.enum) continue;