+++ packages/core/src/rules/oas3/no-unused-components.ts
@@ -2,7 +2,7 @@ import { Oas3Rule } from '../../visitors';
 import { Location } from '../../ref-utils';
 
 export const NoUnusedComponents: Oas3Rule = () => {
-  let components = new Map<string, { used: boolean; location: Location; name: string }>();
+  const components = new Map<string, { used: boolean; location: Location; name: string }>();
 
   function registerComponent(location: Location, name: string): void {
     components.set(location.absolutePointer, {