+++ docs/changelog.md
@@ -5,6 +5,17 @@ tocMaxDepth: 2
 # Redocly CLI changelog
 
 
+## 1.0.0-beta.107 (2022-08-16)
+
+### Changes
+
+- Add `allowAdditionalProperties` to built-in rules and mark `disallowAdditionalProperties` as deprecated
+
+### Fixes
+
+- Introduced severity level `off` for Assertions
+
+
 ## 1.0.0-beta.106 (2022-08-09)
 
 ### Fixes